Consolidate unit test projects
