+++ src/SFA.DAS.Campaign.UnitTests/Application/DataCollection/UserDataCollectionTests/WhenRemovingUserData.cs
