+++ src/SFA.DAS.Campaign.UnitTests/Web/ViewComponents/WhenRenderingTheHeroHeadingViewComponent.cs
