+++ src/SFA.DAS.Campaign.sln
@@ -7,11 +7,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.Campaign.Web", "SFA
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{A89024A5-7D28-42A4-840B-F2F81DF10635}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.Campaign.Web.UnitTests", "SFA.DAS.Campaign.Web.UnitTests\SFA.DAS.Campaign.Web.UnitTests.csproj", "{06D897E8-4472-465F-8499-2508560572AC}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.Campaign.Application", "SFA.DAS.Campaign.Application\SFA.DAS.Campaign.Application.csproj", "{5D7E8906-1680-42CC-A074-1F3B494EAC01}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.Campaign.Application.UnitTests", "SFA.DAS.Campaign.Application.UnitTests\SFA.DAS.Campaign.Application.UnitTests.csproj", "{2C7311D1-3C13-4B39-9B7B-FE44FC229249}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.Campaign.Domain", "SFA.DAS.Campaign.Domain\SFA.DAS.Campaign.Domain.csproj", "{ABA71A27-1FF0-4201-BBC6-E6E1D0ED44A4}"
 EndProject
@@ -20,6 +16,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.Campaign.Infrastructure", "SFA.DAS.Campaign.Infrastructure\SFA.DAS.Campaign.Infrastructure.csproj", "{93CE170F-44C8-4F8E-A749-9154C3555577}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ifa.Api", "Ifa.Api\Ifa.Api.csproj", "{C4F9CCB1-EA2B-49E9-AE9B-14B7C6F46FE9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SFA.DAS.Campaign.UnitTests", "SFA.DAS.Campaign.UnitTests\SFA.DAS.Campaign.UnitTests.csproj", "{06D4B90F-81B1-4551-AFEE-05B0B1104C94}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -31,18 +29,10 @@ Global
 		{78963483-FF2C-42AF-9C29-8DC42B98D6A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{78963483-FF2C-42AF-9C29-8DC42B98D6A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{78963483-FF2C-42AF-9C29-8DC42B98D6A6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{06D897E8-4472-465F-8499-2508560572AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{06D897E8-4472-465F-8499-2508560572AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{06D897E8-4472-465F-8499-2508560572AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{06D897E8-4472-465F-8499-2508560572AC}.Release|Any CPU.Build.0 = Release|Any CPU
 		{5D7E8906-1680-42CC-A074-1F3B494EAC01}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{5D7E8906-1680-42CC-A074-1F3B494EAC01}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5D7E8906-1680-42CC-A074-1F3B494EAC01}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5D7E8906-1680-42CC-A074-1F3B494EAC01}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2C7311D1-3C13-4B39-9B7B-FE44FC229249}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2C7311D1-3C13-4B39-9B7B-FE44FC229249}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2C7311D1-3C13-4B39-9B7B-FE44FC229249}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2C7311D1-3C13-4B39-9B7B-FE44FC229249}.Release|Any CPU.Build.0 = Release|Any CPU
 		{ABA71A27-1FF0-4201-BBC6-E6E1D0ED44A4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{ABA71A27-1FF0-4201-BBC6-E6E1D0ED44A4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ABA71A27-1FF0-4201-BBC6-E6E1D0ED44A4}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -59,13 +49,15 @@ Global
 		{C4F9CCB1-EA2B-49E9-AE9B-14B7C6F46FE9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C4F9CCB1-EA2B-49E9-AE9B-14B7C6F46FE9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C4F9CCB1-EA2B-49E9-AE9B-14B7C6F46FE9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{06D4B90F-81B1-4551-AFEE-05B0B1104C94}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{06D4B90F-81B1-4551-AFEE-05B0B1104C94}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{06D4B90F-81B1-4551-AFEE-05B0B1104C94}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{06D4B90F-81B1-4551-AFEE-05B0B1104C94}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{06D897E8-4472-465F-8499-2508560572AC} = {A89024A5-7D28-42A4-840B-F2F81DF10635}
-		{2C7311D1-3C13-4B39-9B7B-FE44FC229249} = {A89024A5-7D28-42A4-840B-F2F81DF10635}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {DF997E03-26CD-4805-8114-38F71532A883}
